requirements-dev.txt
Django>=4.2.0,<5.0.0
djangorestframework>=3.14.0,<4.0.0
djangorestframework-simplejwt>=5.3.0,<6.0.0
django-cors-headers>=4.3.0,<5.0.0
Pillow>=10.0.0,<11.0.0
python-dotenv>=1.0.0,<2.0.0
numpy>=1.24.0,<2.0.0
scikit-learn>=1.3.0,<2.0.0
psycopg2-binary>=2.9.9,<3.0.0
gunicorn>=21.2.0,<22.0.0
